requirements.txt
Jinja2 >= 3.0.0
Pillow >= 12.0.0